Moved route handling to controller
